AARAMB README.md
I developed a responsive website for College Management under the project titled AARAMB - Where Ideas Meet, using Full Stack Web Development technologies. The website was built to serve as the central platform for a 2-day Bootcamp and a 24-hour hackathon focused on Full Stack Web Development. The event, which took place from 05-04-2023 to 08-04-2023, was organized by Satya Institute of Technology and Management in collaboration with Brainovision Solutions India Pvt. Ltd..